requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.7
Flask==2.2.2
flask-marshmallow==1.2.1
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.15.2
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
numpy==2.1.3
packaging==24.2
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
shapely==2.0.6
SQLAlchemy==1.4.46
typing_extensions==4.12.2
Werkzeug==2.2.3
zipp==3.20.2
